backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
diffusers
transformers
accelerate
safetensors
sentencepiece
Pillow
python-multipart
protobuf
hf-transfer
sqlalchemy
psycopg2-binary
